add test with aer and move tests
