+++ test/python/primitives/test_sampler.py
@@ -1,6 +1,6 @@
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2022.
+# (C) Copyright IBM 2022, 2023.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -12,22 +12,19 @@
 
 """Tests for Sampler."""
 
-import json
 import unittest
 from test import combine
 
 import numpy as np
 from ddt import ddt
 
-from qiskit import QuantumCircuit, pulse, transpile
+from qiskit import QuantumCircuit
 from qiskit.circuit import Parameter
 from qiskit.circuit.library import RealAmplitudes
 from qiskit.exceptions import QiskitError
 from qiskit.extensions.unitary import UnitaryGate
 from qiskit.primitives import Sampler, SamplerResult
-from qiskit.primitives.utils import _circuit_key
 from qiskit.providers import JobStatus, JobV1
-from qiskit.providers.fake_provider import FakeAlmaden
 from qiskit.test import QiskitTestCase
 
 
@@ -744,34 +741,6 @@ class TestSampler(QiskitTestCase):
             self._compare_probs(result.quasi_dists, target)
             self.assertEqual(result.quasi_dists[0].shots, 1024)
 
-    def test_different_circuits(self):
-        """Test collision of quantum circuits."""
-
-        with self.subTest("Ry circuit"):
-
-            def test_func(n):
-                qc = QuantumCircuit(1, 1, name="foo")
-                qc.ry(n, 0)
-                return qc
-
-            keys = [_circuit_key(test_func(i)) for i in range(5)]
-            self.assertEqual(len(keys), len(set(keys)))
-
-        with self.subTest("pulse circuit"):
-
-            def test_with_scheduling(n):
-                custom_gate = pulse.Schedule(name="custom_x_gate")
-                custom_gate.insert(
-                    0, pulse.Play(pulse.Constant(160 * n, 0.1), pulse.DriveChannel(0)), inplace=True
-                )
-                qc = QuantumCircuit(1)
-                qc.x(0)
-                qc.add_calibration("x", qubits=(0,), schedule=custom_gate)
-                return transpile(qc, FakeAlmaden(), scheduling_method="alap")
-
-            keys = [_circuit_key(test_with_scheduling(i)) for i in range(1, 5)]
-            self.assertEqual(len(keys), len(set(keys)))
-
     def test_circuit_with_unitary(self):
         """Test for circuit with unitary gate."""
         gate = UnitaryGate(np.eye(2))
@@ -783,19 +752,6 @@ class TestSampler(QiskitTestCase):
         sampler = Sampler()
         sampler_result = sampler.run([circuit]).result()
         self.assertDictAlmostEqual(sampler_result.quasi_dists[0], {0: 1, 1: 0})
-
-    def test_circuit_key_controlflow(self):
-        """Test for a circuit with control flow."""
-        qc = QuantumCircuit(2, 1)
-
-        with qc.for_loop(range(5)):
-            qc.h(0)
-            qc.cx(0, 1)
-            qc.measure(0, 0)
-            qc.break_loop().c_if(0, True)
-
-        self.assertIsInstance(hash(_circuit_key(qc)), int)
-        self.assertIsInstance(json.dumps(_circuit_key(qc)), str)
 
 
 if __name__ == "__main__":
